requirements.txt
Django==5.2.5
djangorestframework==3.16.1
Pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-environ==0.12.0
openpyxl==3.1.2
pandas==2.2.2
xlsxwriter==3.2.0
django-allauth==0.62.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
